Split babel.config.js across multiple files
