+++ build-system/babel-config/OWNERS
@@ -1,0 +1,14 @@
+// For an explanation of the OWNERS rules and syntax, see:
+// https://github.com/ampproject/amp-github-apps/blob/master/owners/OWNERS.example
+
+{
+  rules: [
+    {
+      owners: [
+        {name: 'ampproject/wg-infra'},
+        {name: 'ampproject/wg-runtime'},
+        {name: 'ampproject/wg-performance'},
+      ],
+    },
+  ],
+}
